requirements.txt
python-binance
pandas
ta
requests
python-dotenv
pywhatkit
flask
